Replace (*frame_iter).get_timestamp() with a template function
This change removes the need for the type stored in the latency buffer
to have a get_timestamp() member function. All of the "frame" types
have such a member function, but the DS object types (TP, TA, TC)
don't, and it doesn't really make sense to add a function: the DS
objects have multiple associated timestamps, and so we don't privilege
one of them in particular in the object.

Instead, we have a template function get_frame_iterator_timestamp(T
iter) which takes the type returned by the begin() and end() functions
in a ReadoutType object and returns whatever timestamp should be used
to determine whether the item pointed to by the iterator is within the
readout window. An implementation that uses get_timestamp() (for the
"frame" ReadoutTypes) is provided in this commit; the DS objects have
template specializations in the trigger package.