At run stop, allow all waiting requests to be serviced or time out
When the run is stopped, there may still be waiting requests for data
that haven't yet been serviced from the buffer. Previous incorrect
behaviour was to check the running flag and return an empty fragment
if the run had been stopped. New behaviour is to wait for each request
to either be serviced from the buffer, or to timeout by going past m_retry_count.